feat: Remove Slideshow component and refactor BackgroundSlideshow
The Slideshow component was removed from the codebase as it was no longer being used. The BackgroundSlideshow component was refactored to remove any references to the Slideshow component. This improves the codebase by removing unnecessary code and dependencies.

Note: This message assumes that the Slideshow component was previously unused and the BackgroundSlideshow component was modified to remove any references to it.